Don't drop slot's time_updated because we didn't make it!

+++ backend/src/appointment/migrations/versions/2024_01_10_2259-502c76bc79e0_add_time_created_and_time_.py
@@ -34,7 +34,8 @@ def upgrade() -> None:
 def downgrade() -> None:
     for table in affected_tables:
         op.drop_column(table, 'time_created')
-        op.drop_column(table, 'time_updated')
+        if table != 'slots':
+            op.drop_column(table, 'time_updated')
 
     for table in index_tables:
         op.drop_index('ix_time_created', table)
